requirements.txt
Django==2.2.2
django-crispy-forms==1.9.0
django-filter==2.2.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2